scripts/requirements.txt
requests>=2.28
google-analytics-data>=0.18
google-cloud-bigquery>=3.10
google-api-python-client>=2.80
google-auth>=2.20
